backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlalchemy>=1.4
asyncpg>=0.27
python-dotenv>=1.0
pydantic>=1.10
httpx>=0.24
openai>=0.28       # optional, used if you choose OpenAI
slowapi>=0.1.4     # rate limiting
pyjwt>=2.8
pytest>=7.4
pytest-asyncio>=0.21
httpx>=0.24